requirements.txt
keyboard==0.13.5
pynput==1.7.7
PyYAML==6.0.2
six==1.16.0